requirements.txt
autopep8==1.6.0
blinker==1.4
boto3==1.19.4
botocore==1.22.4
click==8.0.3
Flask==2.0.2
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.1
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.26
toml==0.10.2
urllib3==1.26.7
uuid==1.30
Werkzeug==2.0.2
WTForms==2.3.3
